Add libcares missing file patch
