+++ src/SPC/builder/unix/library/libcares.php
@@ -5,9 +5,20 @@ declare(strict_types=1);
 namespace SPC\builder\unix\library;
 
 use SPC\exception\RuntimeException;
+use SPC\store\FileSystem;
 
 trait libcares
 {
+    public function patchBeforeBuild(): bool
+    {
+        if (!file_exists($this->source_dir . '/src/lib/thirdparty/apple/dnsinfo.h')) {
+            FileSystem::createDir($this->source_dir . '/src/lib/thirdparty/apple');
+            copy(ROOT_DIR . '/src/globals/extra/libcares_dnsinfo.h', $this->source_dir . '/src/lib/thirdparty/apple/dnsinfo.h');
+            return true;
+        }
+        return false;
+    }
+
     /**
      * @throws RuntimeException
      */
